Fix tshark path for mac and *nix
On macOS, we expect users to have Wireshark installed, not the
standalone tshark utility. When Wireshark is correctly installed, it
doesn't expose tshark directly to the system path. Therefore, we need to
determine the path to tshark manually.

Reverting bc907e2 to fix the script for macOS.
Keeping linux path discovery and extending it to *nix.